The looping variable needed to be able to reach the maximum of the check variable. I had to make the looping variable a u_long instead of a int.

+++ communication/atRTIInterface.c++
@@ -342,7 +342,7 @@ void atRTIInterface::processAmbDiscover(rti13::ObjectHandle obj,
 
 void atRTIInterface::processAmbRemove(rti13::ObjectHandle obj)
 {
-   int                      i;
+   u_long                   i;
    u_long                   c;
    atClassID                classID;
    atKeyedBufferHandler *   buf;
